+++ src/debug/BoxlangDebugDescriptor.ts
@@ -1,10 +1,10 @@
 import * as vscode from "vscode";
-import { BoxLang } from "../utils/BoxLang";
+import { boxLangLauncher } from "../utils/workspaceSetup";
 
 
 export class BoxLangDebugAdapter implements vscode.DebugAdapterDescriptorFactory {
     async createDebugAdapterDescriptor(session: vscode.DebugSession, _executable: vscode.DebugAdapterExecutable | undefined): Promise<vscode.DebugAdapterDescriptor> {
-        const port = await BoxLang.startDebugger();
+        const port = await boxLangLauncher.startDebugger();
 
         return new vscode.DebugAdapterServer(Number.parseInt(port), "localhost");
     }