+++ src/utils/LanguageServer.ts
@@ -1,11 +1,16 @@
 import net from "net";
 import vscode from "vscode";
 import { LanguageClient, ServerOptions } from "vscode-languageclient/node";
-import { BoxLang } from "./BoxLang";
 import { boxlangOutputChannel } from "./OutputChannels";
+import { boxLangLauncher } from "./workspaceSetup";
 
 
 let client: LanguageClient;
+
+export async function restart(){
+    await stop();
+    startLSP();
+}
 
 export async function stop() {
     if (!client) {
@@ -51,7 +56,7 @@ export function getLSPServerConfig(): ServerOptions {
     }
 
     return async () => {
-        const [_process, port] = await BoxLang.startLSP();
+        const [_process, port] = await boxLangLauncher.startLSP();
 
         let socket = net.connect(port, "0.0.0.0");
         return {